fix(GRO-2876): merge expenditures of the same type into one

+++ src/main/java/com/selina/lending/api/dto/common/ExpenditureDto.java
@@ -31,7 +31,7 @@ public class ExpenditureDto {
     @Schema(implementation = Frequency.class)
     @EnumValue(enumClass = Frequency.class)
     String frequency;
-    Integer balanceDeclared;
+    Double balanceDeclared;
 
     @NotNull
     Double amountDeclared;
